upgrades kotlin version to latest and includes the jvmTarget property. this fixes problems with generated code releases. Also workspace samples has been tidied up.
