+++ pom.xml
@@ -21,7 +21,8 @@
 
     <properties>
         <tornadofx.version>1.7.10</tornadofx.version>
-        <kotlin.version>1.1.0</kotlin.version>
+        <kotlin.version>1.1.4</kotlin.version>
+	<kotlin.compiler.jvmTarget>1.8</kotlin.compiler.jvmTarget>
     </properties>
 
     <dependencies>
