+++ workspace/src/main/kotlin/no/tornado/fxsample/workspace/DemoWorkspace.kt
@@ -1,14 +1,10 @@
 package no.tornado.fxsample.workspace
 
 import javafx.application.Platform
-import javafx.beans.binding.BooleanExpression
 import javafx.beans.property.SimpleBooleanProperty
-import javafx.beans.property.StringProperty
-import javafx.scene.Parent
 import javafx.scene.control.Menu
 import javafx.scene.control.MenuItem
 import javafx.scene.control.TextArea
-import javafx.scene.layout.Pane
 import tornadofx.*
 import java.io.IOException
 import java.io.OutputStream
@@ -16,6 +12,9 @@ import java.io.PrintStream
 import java.nio.charset.Charset
 import java.util.*
 
+/**
+ * Created by miguelius on 04/09/2017.
+ */
 class DemoWorkspace : Workspace("Editor") {
     val editorController: EditorController by inject()
 
@@ -87,57 +86,4 @@ class DemoWorkspace : Workspace("Editor") {
         }
     }
 
-}
-
-
-class TextEditorFragment(val documentViewModel: DocumentViewModel) : Fragment(){
-    override val root = pane {
-        title = documentViewModel.title.value
-        textarea (documentViewModel.text) {
-            this.prefWidthProperty().bind(this@pane.widthProperty());
-            this.prefHeightProperty().bind(this@pane.heightProperty());
-
-        }
-    }
-
-    init {
-        documentViewModel.title.addListener { w, o, n ->
-            this.title = n
-        }
-    }
-
-    override val deletable = SimpleBooleanProperty(false)
-    override val closeable = SimpleBooleanProperty( true)
-    override val savable = documentViewModel.dirty
-    override val refreshable = documentViewModel.dirty
-
-    override fun onSave() {
-        documentViewModel.commit()
-    }
-
-    override fun onRefresh() {
-        documentViewModel.rollback()
-    }
-}
-
-class TextAreaOutputStream(val textArea: TextArea): OutputStream() {
-
-    @Throws(IOException::class)
-    override fun write(b: Int) {
-        throw UnsupportedOperationException()
-    }
-
-    @Throws(IOException::class)
-    override fun write(b: ByteArray, off: Int, len: Int) {
-        // redirects data to the text area
-        textArea.appendText(String(Arrays.copyOf(b, len), Charset.defaultCharset()))
-        // scrolls the text area to the end of data
-        textArea.scrollTop = java.lang.Double.MAX_VALUE
-    }
-
-}
-
-class EmptyView : View() {
-    val controller: EditorController by inject()
-    override val root = label(controller.quote())
 }
