+++ workspace/src/main/kotlin/no/tornado/fxsample/workspace/models.kt
@@ -1,18 +1,15 @@
 package no.tornado.fxsample.workspace
 
-import javafx.beans.property.SimpleBooleanProperty
-import javafx.beans.property.SimpleStringProperty
 import tornadofx.ItemViewModel
 import tornadofx.observable
 
+/**
+ * Created by miguelius on 04/09/2017.
+ */
 data class Document(var title : String = "", var text: String = "")
 
 class DocumentViewModel(var document: Document = Document()) : ItemViewModel<Document>() {
     val title = bind { document.observable(Document::title) }
     val text = bind { document.observable(Document::text) }
-
-    init {
-        isDirty(SimpleBooleanProperty(true))
-    }
 
 }
